Code
5/12

+++ QLCB/utils.py
@@ -7,7 +7,8 @@ from urllib.request import urlopen, Request
 from pymysql import Date
 from sqlalchemy import func, extract
 
-from models import BookDetails, Tickets, Flights, TicketTypes, Airports, PaymentMethods, Customers, Stopovers, StopoverDetails
+from models import BookDetails, Tickets, Flights, TicketTypes, Airports, PaymentMethods, Employees, Customers, \
+    Stopovers, StopoverDetails, Rules
 from QLCB import db, app
 from flask_login import current_user
 from hashlib import sha256
@@ -100,7 +101,7 @@ def getYear():
 def get_ticket_types():
     return TicketTypes.query.all()
 
-def get_flights(start=None, destination=None, page=None, id=None, flew=False):
+def get_flights(start=None, destination=None, page=None, id=None, flew=False, takeOffTime=None):
     flights = Flights.query
     if id:
         return flights.get(id)
@@ -113,12 +114,20 @@ def get_flights(start=None, destination=None, page=None, id=None, flew=False):
     if destination:
         flights = flights.filter(Flights.idDestinationAirport == destination)
 
+    if takeOffTime:
+        from datetime import datetime
+        takeOffTime = datetime.strptime(takeOffTime, '%Y-%m-%d')
+        flights = flights.filter(extract('day', Flights.takeOffTime) == takeOffTime.day,
+                                 extract('month', Flights.takeOffTime) == takeOffTime.month,
+                                 extract('year', Flights.takeOffTime) == takeOffTime.year)
+
+    flights = flights.order_by(Flights.takeOffTime)
+
     if page:
         size = app.config["PAGE_SIZE"]
         start = (int(page)-1)*size
         end = start + size
         return flights.all()[start:end], len(flights.all())
-
     return flights.all()
 
 def get_airports_name():
@@ -143,43 +152,57 @@ def cart_stats(cart=None):
     }
 
 
-def add_booking(noBusinessClass, noEconomyClass, customer, employee, flight, error):
+def add_booking(noBusinessClass, noEconomyClass, customer, employee, flight, pay_method, orderKey, status):
     import datetime
     bookTime = datetime.datetime.now()
-    pay_method = get_pay_methods(name='Tiền mặt')
     b = BookDetails(noBusinessClass=noBusinessClass, noEconomyClass=noEconomyClass,
                     bookTime=bookTime, customer=customer, employee=employee,
-                    flight=flight, paymentMethod=pay_method)
+                    flight=flight, paymentMethod=pay_method,
+                    orderKey=orderKey, status=status)
     db.session.add(b)
-    billAdd = False
+    if status == 0:
+        try:
+            db.session.commit()
+            return True
+        except:
+            return False
+    else:
+        billAdd = False
+        try:
+            billAdd = True
+            return add_tickets(bookDetail=b)
+        except Exception as ex:
+            db.session.rollback()
+            if billAdd:
+                db.session.delete(b)
+                db.session.commit()
+            print(ex.args)
+            return False
+
+def add_tickets(bookDetail):
     try:
-        billAdd = True
-        for i in range(0, noBusinessClass):
-            t = Tickets(price=flight.priceBusinessClass, idType=1, bookDetail=b)
+        for i in range(0, bookDetail.noBusinessClass):
+            t = Tickets(price=bookDetail.flight.priceBusinessClass, idType=1, bookDetail=bookDetail)
             db.session.add(t)
-        for i in range(0, noEconomyClass):
-            t = Tickets(price=flight.priceEconomyClass, idType=2, bookDetail=b)
+        for i in range(0, bookDetail.noEconomyClass):
+            t = Tickets(price=bookDetail.flight.priceEconomyClass, idType=2, bookDetail=bookDetail)
             db.session.add(t)
         db.session.commit()
         return True
     except Exception as ex:
-        db.session.rollback()
-        if billAdd:
-            db.session.delete(b)
-            db.session.commit()
-        error.append(ex.args)
+        print(ex.args)
         return False
 
-# region payMomo
-def payByMomo(totalPrice, url):
-    domain =  'http://127.0.0.1:5000/'
+
+def payByMomo(totalPrice, domain):
+    # domain = 'http://14.160.134.135:65001/'
     endpoint = "https://test-payment.momo.vn/gw_payment/transactionProcessor"
     partnerCode = "MOMO544Q20211126"
     accessKey = "FoblaCbnWl9gdHeg"
     serectkey = "8QCHW2eoJJWhZU6TJp0L2dKlngawMaP8"
     orderInfo = "Thanh toán vé máy bay "
-    returnUrl = url
-    notifyurl = domain + ''
+    returnUrl = domain + 'momo/return'
+    notifyurl = domain + 'api/momo/notify'
     amount = totalPrice
     orderId = str(uuid.uuid4())
     requestId = str(uuid.uuid4())
@@ -209,39 +232,63 @@ def payByMomo(totalPrice, url):
     f = urlopen(req)
     response = f.read()
     f.close()
-    import pdb
-    pdb.set_trace()
-    return json.loads(response)['payUrl']
+    return json.loads(response)
 
 
-def add_customer(name, phone, password, error):
+def add_customer(name, phone, password, dob=None, gender=None, idNo=None, address=None):
     password = sha256((password + phone).encode('utf-8')).hexdigest()
-    customer = Customers(customerName=name, phone=phone, password=password)
+    customer = Customers(customerName=name, phone=phone, password=password,
+                         dob=dob if dob else None, gender=gender,
+                         idNo=idNo if idNo else None,
+                         address=address if address else None)
     db.session.add(customer)
     try:
         db.session.commit()
         return True
     except Exception as ex:
         db.session.rollback()
-        error.append(ex.args)
+        print(ex.args)
         return False
 
 def get_customers(id=None):
     customers = Customers.query
     if id:
         return customers.get(int(id))
-
     return customers.all()
 
-def edit_customer(id, name, phone, age, idNo, gender, address, avatar, error):
+def edit_customer(id, name, dob, gender, idNo, phone, address, avatar, error):
     customer = get_customers(id)
     customer.customerName = name
-    customer.phone = phone
-    customer.age = age
-    customer.idNo = idNo
+    customer.dob = dob if dob else None
     customer.gender = gender
-    customer.address = address
+    customer.idNo = idNo if idNo else None
+    customer.phone = phone
+    customer.address = address if address else None
     customer.avatar = avatar if avatar else customer.avatar
+    try:
+        db.session.commit()
+    except Exception as ex:
+        error.append(ex.args)
+        db.session.rollback()
+        return False
+    return True
+
+def get_employees(id=None):
+    employees = Employees.query
+    if id:
+        return employees.get(int(id))
+    return employees.all()
+
+def edit_employee(employee, username, employeeName, dob, gender, idNo, email, phone, address, avatar, error):
+    employee.username = username
+    employee.employeeName = employeeName
+    employee.dob = dob if dob else None
+    employee.gender = gender
+    employee.idNo = idNo if idNo else None
+    employee.email = email if email else None
+    employee.phone = phone if phone else None
+    employee.address = address if address else None
+    employee.avatar = avatar if avatar else employee.avatar
     try:
         db.session.commit()
     except Exception as ex:
@@ -256,28 +303,80 @@ def get_pay_methods(name=None):
         return pm.filter(PaymentMethods.PMethodName == name).first()
     return pm.all()
 
-def get_book_detail(id):
-    return BookDetails.query.get(id)
+def get_book_detail(id = None, orderKey = None, cid= None):
+    book_details = BookDetails.query
+    if id:
+        return BookDetails.query.get(id)
+    if orderKey:
+        return book_details.filter(BookDetails.orderKey == orderKey).first()
+    if cid:
+        book_details = book_details.filter(BookDetails.idCustomer == cid)
+    return book_details.all()
 
 def get_stopover_detail(fid = None):
     sd = StopoverDetails.query
     if fid:
         sd = sd.filter(StopoverDetails.idFlight == fid)
-
     return sd.all()
-
 
 def get_tickets(cid=None):
     tickets = Tickets.query
     if cid:
         tickets = tickets.filter(Tickets.id)
+    return tickets.all()
 
 def get_slot_remain(fid):
     b, e = db.session.query(func.sum(BookDetails.noBusinessClass),
-                         func.sum(BookDetails.noEconomyClass)).filter(
-        BookDetails.idFlight == fid).one()
+                         func.sum(BookDetails.noEconomyClass))\
+        .filter(BookDetails.idFlight == fid, BookDetails.status == 1).one()
     return {
         "economy": e if e else 0,
         "business": b if b else 0
     }
 
+def paid_book_detail(book_detail):
+    try:
+        book_detail.status = 1
+        db.session.add(book_detail)
+        db.session.commit()
+        return True
+    except Exception as ex:
+        print(ex.args)
+        return False
+
+def add_ticket_type(name):
+    ticket_type = TicketTypes(typeName=name)
+    db.session.add(ticket_type)
+    try:
+        db.session.commit()
+    except Exception as ex:
+        print(ex.args)
+
+def customer_change_password(customer, newpwd):
+    newpwd = sha256((newpwd + customer.phone).encode('utf-8')).hexdigest()
+    customer.password = newpwd
+    db.session.add(customer)
+    try:
+        db.session.commit()
+        return True
+    except Exception as ex:
+        print(ex.args)
+        return False
+
+def employee_change_password(employee, newpwd):
+    newpwd = sha256((newpwd + employee.username).encode('utf-8')).hexdigest()
+    employee.password = newpwd
+    db.session.add(employee)
+    try:
+        db.session.commit()
+        return True
+    except Exception as ex:
+        print(ex.args)
+        return False
+
+def get_rules(id=None, name=None):
+    rules = Rules.query
+    if name:
+        return rules.filter(Rules.ruleName == name).one()
+    if id:
+        return rules.get(id)
